moving code from controller to service
